+++ app/src/hooks/index.ts
@@ -5,3 +5,4 @@
 
 export { useAdminAuth } from "./useAdminAuth";
 export { useFetch, useApiMutation } from "./useFetch";
+export { useTerminalAdmin } from "./useTerminalAdmin";